pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bnote"
version = "3.0.4"
description = "bnote application"
authors = [{name = "Eurobraille team"}]
requires-python = ">=3.11, <3.12"
dependencies= [
    "pyserial ==3.5, <3.6",
    "rpi-gpio ==0.7.1",
    "python-vlc ==3.0.20123",
    "xlrd3 ==1.1.0",
    "striprtf ==0.0.26",
    "pyrtitions ==0.2.3",
    "unidecode ==1.3.8",
    "ebooklib ==0.18",
    "bs4 ==0.0.2",
    "mutagen ==1.47.0",
    "mido ==1.3.2",
    "requests ==2.31.0",
    "pbkdf2 ==1.3",
    "requests-oauthlib ==2.0.0",
    "cryptography ==42.0.7",
    "py-machineid ==0.5.1",
    "wikipedia ==1.4.0",
    "dbus-python ==1.3.2",
    "sdbus-networkmanager ==2.0.0",
    "sdbus ==0.11.1",
    "pyproject_hooks ==1.1.0",
    "build ==1.2.1",
    "translate ==3.6.1"
    ]

[tool.setuptools.package-dir]
bnote = "bnote"
